requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
langchain-groq>=0.1.0
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-community>=0.0.1
faiss-cpu>=1.7.0
requests>=2.26.0
numpy<2.0.0
pandas>=1.3.0
scikit-learn>=0.24.0
sentence-transformers>=2.2.0
torch>=2.0.0
python-dateutil